Moved get_class_from string to utils
